feat: Add hot reloading to bot & move types folder into src
