+++ src/types/tic.d.ts
